# Conflicts:
#	include/vrv/doc.h
#	src/doc.cpp

+++ include/vrv/note.h
@@ -176,10 +176,10 @@ public:
     void SetScoreTimeTiedDuration(double timeInSeconds);
     void SetMIDIPitch(char pitch);
     double GetScoreTimeOnset();
-    int GetRealTimeOnsetMilliseconds();
+    double GetRealTimeOnsetMilliseconds();
     double GetScoreTimeOffset();
     double GetScoreTimeTiedDuration();
-    int GetRealTimeOffsetMilliseconds();
+    double GetRealTimeOffsetMilliseconds();
     double GetScoreTimeDuration();
     char GetMIDIPitch();
     ///@}
@@ -306,14 +306,14 @@ private:
     /**
      * The time in milliseconds since the start of the measure element that contains the note.
      */
-    int m_realTimeOnsetMilliseconds;
+    double m_realTimeOnsetMilliseconds;
 
     /**
      * The time in milliseconds since the start of the measure element to end of printed note.
      * The real-time duration of a tied group is not currently tracked (this gets complicated
      * if there is a tempo change during a note sustain, which is currently not supported).
      */
-    int m_realTimeOffsetMilliseconds;
+    double m_realTimeOffsetMilliseconds;
 
     /**
      * If the note is the first in a tied group, then m_scoreTimeTiedDuration contains the
